Seperate file to remove table from the one creating it
