+++ .travis.yml
@@ -10,7 +10,7 @@ services:
   - postgresql
 before_script:
   - psql -c 'create database foodfast_test;' -U postgres
-  - npm run createTable
+  - npm run table
 before_install:
   - npm install coveralls
 after_script:
