+++ db/dropTable.js
@@ -1,0 +1,13 @@
+import db from './index';
+
+db.query('DROP TABLE IF EXISTS orders', (err) => {
+  if (err) {
+    console.log(err);
+  }
+});
+
+db.query('DROP TABLE IF EXISTS foodlist', (err) => {
+  if (err) {
+    console.log(err);
+  }
+});
