+++ db/table.js
@@ -1,22 +1,4 @@
 import db from './index';
-// create table query
-db.query('DROP TABLE IF EXISTS foodlist, orders, users', (err) => {
-  if (err) {
-    console.log(err);
-  }
-});
-
-
-db.query(`create table foodlist(
-id serial PRIMARY KEY,
-food text NOT NULL,
-price numeric NOT NULL,
-image text NOT NULL
-)`, (err) => {
-  if (err) {
-    console.log(err);
-  }
-});
 
 db.query(`create table orders(
 id serial PRIMARY KEY,
@@ -25,6 +7,17 @@ quantity text NOT NULL,
 price numeric NOT NULL,
 user_id integer NOT NULL,
 status status NOT NULL
+)`, (err) => {
+  if (err) {
+    console.log(err);
+  }
+});
+
+db.query(`CREATE TABLE foodlist(
+id serial PRIMARY KEY,
+food text NOT NULL,
+price numeric NOT NULL,
+image text NOT NULL
 )`, (err) => {
   if (err) {
     console.log(err);
