+++ package.json
@@ -5,10 +5,12 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint ./routes --fix",
-    "test": "nyc mocha ./dist/test/appTest.js --exit",
+    "test": "nyc mocha ./dist/test --exit",
     "build": "rm -rf dist/ && babel ./ --out-dir dist/ --ignore ./node_modules,./.babelrc,./package.json,./npm-debug.log",
     "start": "npm run build && nodemon dist/index.js",
-    "createTable": "node dist/db/table.js",
+    "drop_table": "node dist/db/dropTable.js",
+    "create_table": "node dist/db/table.js",
+    "table": "npm run drop_table && npm run create_table",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
